Start linking something
- Check we can find the package
- Rename package
- Symlink package

+++ src/Garoevans/ComposerSym/ComposerSym.php
@@ -41,41 +41,67 @@ class ComposerSym extends CliCommand
     $this->composerJson = ComposerJson::get($this->projectDir);
     $this->setHomeDir($this->homeDir);
 
-    printf("\n> Starting to process composer packages.\n\n");
+    printf("\n> Starting to process composer packages.\n");
 
     $composerJsonObj = $this->composerJson->getParsedComposerJsonFile();
     foreach ($composerJsonObj->require as $package => $version) {
-      $doSymlink = UserPrompt::confirm(
-        sprintf("> Would you like to symlink %s?", $package),
-        'y'
-      );
 
-      if ($doSymlink) {
-        $linkToPotentialLocation = false;
-        $potentialLocation = build_path($this->homeDir, $package);
-        if (file_exists($potentialLocation)) {
-          $linkToPotentialLocation = UserPrompt::confirm(
-            sprintf("> Link to '%s'?", $potentialLocation)
-          );
-        }
+      // TODO: move into ComposerVendor class
+      if (file_exists(build_path($this->projectDir, "vendor", $package))) {
+        $doSymlink = UserPrompt::confirm(
+          sprintf("\n> Would you like to symlink %s?", $package),
+          'y'
+        );
 
-        if (!$linkToPotentialLocation) {
-          do {
-            if (isset($linkTo)) {
-              printf("> Directory '%s' does not exist.\n", $linkTo);
-            }
-            $linkTo = UserPrompt::prompt(
-              sprintf("> Enter full path to base directory for '%s'", $package)
+        if ($doSymlink) {
+          // TODO: move to method
+          $linkToPotentialLocation = false;
+          $potentialLocation = build_path($this->homeDir, $package);
+          if (file_exists($potentialLocation)) {
+            $linkToPotentialLocation = UserPrompt::confirm(
+              sprintf("> Link to '%s'?", $potentialLocation)
             );
-          } while (!file_exists($linkTo));
-        } else {
-          $linkTo = $linkToPotentialLocation;
+          }
+
+          // TODO: move to method
+          if (!$linkToPotentialLocation) {
+            do {
+              if (isset($linkTo)) {
+                printf("> Directory '%s' does not exist.\n", $linkTo);
+              }
+              $linkTo = UserPrompt::prompt(
+                sprintf(
+                  "> Enter full path to base directory for '%s': ",
+                  $package
+                )
+              );
+            } while (!file_exists($linkTo));
+          } else {
+            $linkTo = $linkToPotentialLocation;
+          }
+
+          $packageParts = explode("/", $package);
+          end($packageParts);
+          $packageParts[key($packageParts)] = "__" . current($packageParts);
+          $newPackage = implode("/", $packageParts);
+
+          $oldPath = build_path($this->projectDir, $this->vendor, $package);
+          $newPath = build_path($this->projectDir, $this->vendor, $newPackage);
+
+
+          rename($oldPath, $newPath);
+          symlink($linkTo, $oldPath);
+
+          sprintf("> %s symlinked:", $package);
+          sprintf(">> link: %s", $oldPath);
+          sprintf(">> target: %s", $linkTo);
+
+          // TODO: Write to log file in user dir ~
+
+
+          unset($linkTo);
+          echo "\n";
         }
-
-        // TODO: Symlink location
-        // TODO: Write to log file in user dir ~
-
-        echo "\n";
       }
     }
   }
